requirements.txt
streamlit==1.28.0
chromadb==0.4.15
sentence-transformers==2.2.2
google-generativeai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.0
pypdf2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
urllib3==1.26.0
huggingface-hub==0.16.4
numpy==1.24.3